Splitted Application codes into different files for better understanding.
* All codes moved into `blinkeye/`
* `main.py` for starting the application
* `blinkeye/__init__.py` to let the 'blinkeye' folder work like a package
* `blinkeye/audio.py` for Audio related code
* `blinkeye/dashboard.py` for Dashboard's code
* `blinkeye/notifier.py` for Blink Eye Notifier's code
* `blinkeye/utils.py` for utility functions and helper code
